requirements.txt
Flask>=2.0
torch>=1.8           # If your OCR is a PyTorch model
torchvision>=0.9     # If using torchvision transforms
opencv-python>=4.5   # For preprocessing
Pillow>=8.2
pytesseract>=0.3     # If you choose Tesseract for OCR
sympy>=1.8
gunicorn>=20.1       # For production server
flask-cors>=3.0      # If you need CORS
